voice-automation-hub/backend/requirements.txt
# Core Dependencies
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
pydantic-settings>=2.5.2

# OpenAI and ChatKit
openai>=1.0.0
openai-chatkit>=1.0.0
openai-agents>=0.3.2

# HTTP and Networking
httpx>=0.27.0
python-multipart>=0.0.9

# Configuration
python-dotenv>=1.0.1

# Storage
aiosqlite>=0.20.0

# Authentication & Security
PyJWT>=2.8.0
bcrypt>=4.1.2

# Caching (optional - Redis)
redis>=5.0.1

# WebSocket
websockets>=12.0

# Development and Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=5.0.0
ruff>=0.6.0
mypy>=1.11.0
